+++ lib/contentful/management/http_client.rb
@@ -9,8 +9,8 @@ module Contentful
       # @param [Hash] headers
       #
       # @return [HTTP::Response]
-      def get_http(url, query, headers = {})
-        HTTP[headers].get(url, params: query)
+      def get_http(url, query, headers = {}, proxy = {})
+        http_send(:get, url, { params: query }, headers, proxy)
       end
 
       # Post Request
@@ -20,8 +20,8 @@ module Contentful
       # @param [Hash] headers
       #
       # @return [HTTP::Response]
-      def post_http(url, params, headers = {})
-        HTTP[headers].post(url, json: params)
+      def post_http(url, params, headers = {}, proxy = {})
+        http_send(:post, url, { json: params }, headers, proxy)
       end
 
       # Delete Request
@@ -31,8 +31,8 @@ module Contentful
       # @param [Hash] headers
       #
       # @return [HTTP::Response]
-      def delete_http(url, params, headers = {})
-        HTTP[headers].delete(url, params: params)
+      def delete_http(url, params, headers = {}, proxy = {})
+        http_send(:delete, url, { params: params }, headers, proxy)
       end
 
       # Put Request
@@ -42,8 +42,41 @@ module Contentful
       # @param [Hash] headers
       #
       # @return [HTTP::Response]
-      def put_http(url, params, headers = {})
-        HTTP[headers].put(url, json: params)
+      def put_http(url, params, headers = {}, proxy = {})
+        http_send(:put, url, { json: params }, headers, proxy)
+      end
+
+      # Proxy Helper
+      #
+      # @param [Symbol] type
+      # @param [String] url
+      # @param [Hash] params
+      # @param [Hash] headers
+      # @param [Hash] proxy
+      #
+      # @return [HTTP::Response]
+      def proxy_send(type, url, params, headers, proxy)
+        HTTP[headers].via(
+          proxy[:host],
+          proxy[:port],
+          proxy[:username],
+          proxy[:password]
+        ).public_send(type, url, params)
+      end
+
+      # HTTP Helper
+      # Abtracts the Proxy/No-Proxy logic
+      #
+      # @param [Symbol] type
+      # @param [String] url
+      # @param [Hash] params
+      # @param [Hash] headers
+      # @param [Hash] proxy
+      #
+      # @return [HTTP::Response]
+      def http_send(type, url, params, headers, proxy)
+        return proxy_send(type, url, params, headers, proxy) unless proxy[:host].nil?
+        HTTP[headers].public_send(type, url, params)
       end
     end
   end